ai-search-web/pyproject.toml
[project]
name = "ai-search-web"
version = "0.1.0"
description = "Streamlit UI for browsing ai-search reports"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv>=1.1.1",
    "streamlit>=1.49.1",
    "elasticsearch>=8.17.0",
]
